service/webrtc/requirements.txt
openai
fastapi
python-dotenv
elevenlabs
git+https://github.com/ai-poet/fastrtc.git#egg=fastrtc[vad,stt,tts]
requests
twilio
mem0ai
litellm
humaware-vad